skill-flow/data/skills-refined-autoskill-36k/skillsmp/gemini-image-generator-1/scripts/requirements.txt
google-genai
pillow
requests
